LayoutTests/fast/forms/ValidityState-patternMismatch-expected.txt

This test checks validity.patternMismatch.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS patternMismatchFor("simple") is false
PASS patternMismatchFor("no-pattern-and-no-value") is false
PASS patternMismatchFor("ip-address") is false
PASS patternMismatchFor("email-address") is false
PASS patternMismatchFor("wrong-email-address") is true
PASS patternMismatchFor("match-1") is false
PASS patternMismatchFor("match-2") is false
PASS patternMismatchFor("match-3") is false
PASS patternMismatchFor("match-4") is false
PASS patternMismatchFor("match-5") is false
PASS patternMismatchFor("match-6") is false
PASS patternMismatchFor("match-7") is false
PASS patternMismatchFor("match-8") is false
PASS patternMismatchFor("match-9") is false
PASS patternMismatchFor("match-10") is false
PASS patternMismatchFor("match-11") is false
PASS patternMismatchFor("match-12") is false
PASS patternMismatchFor("match-13") is false
PASS patternMismatchFor("match-14") is false
PASS patternMismatchFor("match-15") is false
PASS patternMismatchFor("match-16") is false
PASS patternMismatchFor("match-17") is false
PASS patternMismatchFor("match-18") is false
PASS patternMismatchFor("match-19") is false
PASS patternMismatchFor("wrong-gray-or-grey") is true
PASS patternMismatchFor("gray") is false
PASS patternMismatchFor("grey") is false
PASS patternMismatchFor("empty-gray-or-grey") is false
PASS patternMismatchFor("mismatch-1") is true
PASS patternMismatchFor("mismatch-2") is true
PASS patternMismatchFor("mismatch-3") is true
PASS patternMismatchFor("mismatch-4") is true
PASS patternMismatchFor("mismatch-5") is true
PASS patternMismatchFor("mismatch-6") is true
PASS patternMismatchFor("mismatch-7") is true
PASS patternMismatchFor("mismatch-8") is true
PASS patternMismatchFor("mismatch-9") is true
PASS patternMismatchFor("mismatch-10") is true
PASS patternMismatchFor("mismatch-11") is true
PASS patternMismatchFor("mismatch-12") is true
PASS patternMismatchFor("mismatch-13") is true
PASS patternMismatchFor("mismatch-14") is true
PASS patternMismatchFor("mismatch-15") is true
PASS patternMismatchFor("mismatch-16") is true
PASS patternMismatchFor("mismatch-17") is true
PASS patternMismatchFor("mismatch-18") is true
PASS patternMismatchFor("mismatch-19") is true
PASS patternMismatchFor("mismatch-20") is true
PASS patternMismatchFor("empty-pattern-match") is false
PASS patternMismatchFor("empty-pattern-mismatch") is true
PASS patternMismatchFor("disabled") is false
PASS successfullyParsed is true

TEST COMPLETE
